\name{EstimationMethod-class}
\Rdversion{1.1}
\docType{class}
\alias{EstimationMethod}
\alias{EstimationMethod-class}

\title{Class "EstimationMethod"  }
\description{	 This is a class union formed by the classes CvSettings,
  McSettings, HldSettings, LoocvSettings and BootSettings}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "EstimationMethod" in the signature.
}
\references{ Torgo, L. (2013) \emph{An Infra-Structure for Performance
    Estimation and Experimental Comparison of Predictive Models}.
  \url{https://github.com/ltorgo/performanceEstimation}  
}
\author{ Luis Torgo \email{ltorgo@dcc.fc.up.pt} }
\seealso{
  \code{\linkS4class{CV}},
  \code{\linkS4class{MonteCarlo}},
  \code{\linkS4class{LOOCV}},
  \code{\linkS4class{Holdout}},
  \code{\linkS4class{Bootstrap}},
   \code{\linkS4class{EstimationTask}}
}
\examples{
showClass("EstimationMethod")
}
\keyword{classes}
