\name{EstimationSettings-class}
\Rdversion{1.1}
\docType{class}
\alias{EstimationSettings}
\alias{EstimationSettings-class}

\title{Class "EstimationSettings"  }
\description{	 This is a class union formed by the classes CvSettings,
  McSettings, HldSettings, LoocvSettings and BootSettings}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "EstimationSettings" in the signature.
}
\references{ Torgo, L. (2013) \emph{An Infra-Structure for Performance
    Estimation and Experimental Comparison of Predictive Models}.
  \url{https://github.com/ltorgo/performanceEstimation}  
}
\author{ Luis Torgo \email{ltorgo@dcc.fc.up.pt} }
\seealso{
  \code{\linkS4class{CvSettings}},
  \code{\linkS4class{McSettings}},
  \code{\linkS4class{LoocvSettings}},
  \code{\linkS4class{HldSettings}},
  \code{\linkS4class{BootSettings}},
   \code{\linkS4class{EstimationTask}}
}
\examples{
showClass("EstimationSettings")
}
\keyword{classes}
